requirements.txt
psycopg2-binary==2.9.12
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
pyodbc>=4.0.0
